salt/utils/minions.py: Added Access Control in mine_get.
Detection on whether a mine-entry actually has AC is done using the string MINE_MAGIC_DATA_ID. If the mine-entry is a dict and has this key, then it supports AC.
If it is not a dict, or does not have the key, fall back to the old behaviour instead.
In case of AC present, check if the "allow_tgt"-key is present, match the requesting minion against the expanded tgt and return data if it matches.